Add files via upload
Program plays Tic-tac-toe with a user. Each round, the game prints out the state of the board, asks the user where they would like to place their mark, and implements this decision. The program then places its own mark on a randomly chosen available position. Once one of the players won, the program declares the result and asks if the user would like to continue. The first player is selected at random. The description of the game can be found at https://en.wikipedia.org/wiki/Tic-tac-toe.